# trial comment  작업

+++ src/main/java/com/kr/lg/module/board/service/impl/BoardServiceImpl.java
@@ -7,6 +7,8 @@ import com.kr.lg.db.repositories.BoardAttachRepository;
 import com.kr.lg.db.repositories.BoardRecommendRepository;
 import com.kr.lg.db.repositories.BoardRepository;
 import com.kr.lg.enums.*;
+import com.kr.lg.module.board.model.event.BoardRecommendEvent;
+import com.kr.lg.module.board.model.event.UserBoardCreateCountEvent;
 import com.kr.lg.module.board.model.req.*;
 import com.kr.lg.module.board.model.dto.BoardReportDto;
 import com.kr.lg.module.board.model.dto.BoardEnrollDto;
@@ -24,6 +26,7 @@ import com.kr.lg.web.dto.mapper.BoardParam;
 import com.kr.lg.module.board.model.mapper.FindBoardsParamData;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.*;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -54,6 +57,7 @@ public class BoardServiceImpl implements BoardService {
     private final CommentMapper commentMapper;
 
     private final BCryptPasswordEncoder encoder;
+    private final ApplicationEventPublisher applicationEventPublisher;
 
     /**
      * 포지션 게시판 조회 리스트
@@ -63,7 +67,6 @@ public class BoardServiceImpl implements BoardService {
      */
     @Override
     public Page<BoardEntry> findBoards(FindBoardRequest request) throws BoardException {
-        log.info("▶ [포지션 게시판] findBoards 메소드 실행");
         Pageable pageable = PageRequest.of(request.getPage(), request.getPageNum(), getSort(request.getTopic())); // pageable 생성
         MapperParam param = FindBoardsParamData.builder()
                 .type(request.getType())
@@ -83,7 +86,6 @@ public class BoardServiceImpl implements BoardService {
      */
     @Override
     public Page<BoardEntry> findMyBoards(FindMyBoardRequest request, UserTb userTb) throws BoardException {
-        log.info("▶ [포지션 게시판] findMyBoards 메소드 실행");
         Pageable pageable = PageRequest.of(request.getPage(), request.getPageNum(), getSort(request.getTopic())); // pageable 생성
         MapperParam param = FindBoardsParamData.builder()
                 .type(request.getType())
@@ -103,7 +105,6 @@ public class BoardServiceImpl implements BoardService {
      */
     @Override
     public Page<BoardEntry> findLawFirmBoards(FindLawFirmBoardRequest request) throws BoardException {
-        log.info("▶ [포지션 게시판] findLawFirmBoards 메소드 실행");
         Pageable pageable = PageRequest.of(request.getPage(), request.getPageNum(), getSort(request.getTopic())); // pageable 생성
         MapperParam param = FindBoardsParamData.builder()
                 .subject(request.getSubject())
@@ -122,7 +123,6 @@ public class BoardServiceImpl implements BoardService {
      */
     @Override
     public BoardEntry findBoardWithNotLogin(long boardId) throws BoardException {
-        log.info("▶ [포지션 게시판] findBoardIdWithNotLogin 메소드 실행");
         MapperParam param = FindBoardParamData.builder()
                 .boardId(boardId)
                 .build();
@@ -140,7 +140,6 @@ public class BoardServiceImpl implements BoardService {
      */
     @Override
     public BoardEntry findBoardWithLogin(long boardId, UserTb userTb) throws BoardException {
-        log.info("▶ [포지션 게시판] findBoardIdWithLogin 메소드 실행");
         MapperParam param = FindBoardParamData.builder()
                 .boardId(boardId)
                 .userId(userTb.getUserId())
@@ -159,7 +158,6 @@ public class BoardServiceImpl implements BoardService {
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void enrollBoardWithNotLogin(EnrollBoardWithNotLoginRequest request, String ip) throws BoardException {
-        log.info("▶ [포지션 게시판] enrollBoardWithNotLogin 메소드 실행");
         boolean isEnrollFile = request.getFiles() != null && !request.getFiles().isEmpty();
 
         BoardEnrollDto board = BoardEnrollDto.builder()
@@ -190,8 +188,6 @@ public class BoardServiceImpl implements BoardService {
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void enrollBoardWithLogin(EnrollBoardWithLoginRequest request, String ip, UserTb userTb) throws BoardException {
-        log.info("▶ [포지션 게시판] enrollBoardWithLogin 메소드 실행");
-
         if (userTb == null) throw new BoardException(BoardResultCode.NOT_EXIST_USER); // 로그인 필수
         boolean isEnrollLawFirm = request.getIsLawFirm() != null && request.getIsLawFirm() == 1;
         boolean isEnrollFile = request.getFiles() != null && !request.getFiles().isEmpty();
@@ -209,6 +205,7 @@ public class BoardServiceImpl implements BoardService {
                 .build();
 
         BoardTb boardTb = boardEnrollService.enrollBoard(board);
+        applicationEventPublisher.publishEvent(new UserBoardCreateCountEvent(userTb, 1));
         if (isEnrollFile) {
             boardEnrollService.enrollBoardFiles(boardTb, request.getFiles());
         }
@@ -224,8 +221,6 @@ public class BoardServiceImpl implements BoardService {
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void enrollBoardWithLawFirmLogin(EnrollBoardWithLawFirmLoginRequest request, String ip, UserTb userTb) throws BoardException {
-        log.info("▶ [포지션 게시판] enrollBoardWithLawFirmLogin 메소드 실행");
-
         if (userTb == null) throw new BoardException(BoardResultCode.NOT_EXIST_USER); // 로그인 필수
         else if (userTb.getLawFirmId() == null) throw new BoardException(BoardResultCode.NOT_EXIST_LAW_FIRM); // 로펌 필수
         else if (!Objects.equals(userTb.getLawFirmId().getLawFirmId(), request.getId())) throw new BoardException(BoardResultCode.UN_MATCHED_LAW_FIRM_USER);
@@ -244,6 +239,7 @@ public class BoardServiceImpl implements BoardService {
                 .build();
 
         BoardTb boardTb = boardEnrollService.enrollBoard(board);
+        applicationEventPublisher.publishEvent(new UserBoardCreateCountEvent(userTb, 1));
         if (isEnrollFile) {
             boardEnrollService.enrollBoardFiles(boardTb, request.getFiles());
         }
@@ -258,8 +254,6 @@ public class BoardServiceImpl implements BoardService {
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void updateBoardWithNotLogin(UpdateBoardWithNotLoginRequest request) throws BoardException {
-        log.info("▶ [포지션 게시판] updateBoardWithNotLogin 메소드 실행");
-
         Optional<BoardTb> boardTb = boardRepository.findByBoardIdAndWriterType(request.getId(), WriterEnum.ANONYMOUS_TYPE);
         if (boardTb.isPresent()) {
             if (!encoder.matches(request.getPassword(), boardTb.get().getPassword())) throw new BoardException(BoardResultCode.UN_MATCH_PASSWORD);
@@ -288,8 +282,6 @@ public class BoardServiceImpl implements BoardService {
     @Override
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void updateBoardWithLogin(UpdateBoardWithLoginRequest request, UserTb userTb) throws BoardException {
-        log.info("▶ [포지션 게시판] updateBoardWithLogin 메소드 실행");
-
         Optional<BoardTb> boardTb = boardRepository.findByBoardIdAndWriterType(request.getId(), WriterEnum.MEMBER_TYPE);
         if (boardTb.isPresent()) {
             if (!userTb.getUserId().equals(boardTb.get().getUserTb().getUserId())) throw new BoardException(BoardResultCode.UN_MATCHED_USER);
@@ -316,8 +308,6 @@ public class BoardServiceImpl implements BoardService {
      */
     @Override
     public void deleteBoardWithNotLogin(DeleteBoardWithNotLoginRequest request) throws BoardException {
-        log.info("▶ [포지션 게시판] deleteBoardWithNotLogin 메소드 실행");
-
         Optional<BoardTb> boardTb = boardRepository.findByBoardIdAndWriterTypeAndStatus(request.getId(), WriterEnum.ANONYMOUS_TYPE, StatusEnum.NORMAL_STATUS);
         if (boardTb.isPresent()) {
             if (!encoder.matches(request.getPassword(), boardTb.get().getPassword())) throw new BoardException(BoardResultCode.UN_MATCH_PASSWORD);
@@ -335,15 +325,15 @@ public class BoardServiceImpl implements BoardService {
      * @throws BoardException
      */
     @Override
+    @Transactional
     public void deleteBoardWithLogin(DeleteBoardWithLoginRequest request, UserTb userTb) throws BoardException {
-        log.info("▶ [포지션 게시판] deleteBoardWithLogin 메소드 실행");
-
         Optional<BoardTb> boardTb = boardRepository.findByBoardIdAndWriterTypeAndStatus(request.getId(), WriterEnum.MEMBER_TYPE, StatusEnum.NORMAL_STATUS);
         if (boardTb.isPresent()) {
             boolean isBestOrRecommendBoard = boardTb.get().getPostType().equals(PostEnum.BEST_TYPE) || boardTb.get().getPostType().equals(PostEnum.RECOMMEND); // 베스트 or 추천 게시판 플래그
             if (!userTb.getUserId().equals(boardTb.get().getUserTb().getUserId())) throw new BoardException(BoardResultCode.UN_MATCHED_USER); // 작성자 체크
             else if (isBestOrRecommendBoard && !encoder.matches(request.getPassword(), boardTb.get().getPassword())) throw new BoardException(BoardResultCode.UN_MATCH_PASSWORD); // 베스트 or 추천 게시판은 패스워드 검증
             boardDeleteService.deleteBoard(boardTb.get().getBoardId());
+            applicationEventPublisher.publishEvent(new UserBoardCreateCountEvent(userTb, -1));
         } else {
             throw new BoardException(BoardResultCode.NOT_EXIST_BOARD); // 게시판 미존재
 
@@ -359,7 +349,6 @@ public class BoardServiceImpl implements BoardService {
     @Override
     @Transactional
     public void reportBoard(ReportBoardRequest request, String ip) throws BoardException {
-        log.info("▶ [포지션 게시판] reportBoard 메소드 실행");
         BoardReportDto reportDto = BoardReportDto.builder()
                 .ip(ip)
                 .content(request.getContent())
@@ -375,12 +364,12 @@ public class BoardServiceImpl implements BoardService {
      * @throws BoardException
      */
     @Override
+    @Transactional
     public void recommendBoard(RecommendBoardRequest request, UserTb userTb) throws BoardException {
-        log.info("▶ [포지션 게시판] recommendBoard 메소드 실행");
-
         Optional<BoardRecommendTb> boardRecommendTb = boardRecommendRepository.findByBoardTb_BoardIdAndUserTb_UserId(request.getId(), userTb.getUserId()); // 나의 추천 내역 조회
         if (boardRecommendTb.isPresent()) throw new BoardException(BoardResultCode.ALREADY_RECOMMEND_BOARD); // 중복 추천 방어코드
         boardRecommendService.recommendBoard(BoardTb.builder().boardId(request.getId()).build(), userTb); // 게시판 추천
+        applicationEventPublisher.publishEvent(new BoardRecommendEvent(request.getId(), 1)); // 추천 수 증가
     }
 
     /**
@@ -391,7 +380,8 @@ public class BoardServiceImpl implements BoardService {
      */
     @Override
     public void deleteRecommendBoard(DeleteRecommendBoardRequest request, UserTb userTb) throws BoardException {
-        log.info("▶ [포지션 게시판] deleteRecommendBoard 메소드 실행");
+        Optional<BoardRecommendTb> boardRecommendTb = boardRecommendRepository.findByBoardTb_BoardIdAndUserTb_UserId(request.getId(), userTb.getUserId()); // 나의 추천 내역 조회
+        if (!boardRecommendTb.isPresent()) throw new BoardException(BoardResultCode.ALREADY_DELETE_RECOMMEND_BOARD); // 이미 추천 취소 처리
         boardRecommendService.deleteRecommendBoard(request.getId(), userTb.getUserId()); // 게시판 추천 취소
     }
 
@@ -402,8 +392,6 @@ public class BoardServiceImpl implements BoardService {
      */
     @Override
     public void loginBoardWithNotLogin(LoginBoardWithNotLoginRequest request) throws BoardException {
-        log.info("▶ [포지션 게시판] loginBoardWithNotLogin 메소드 실행");
-
         Optional<BoardTb> boardTb = boardRepository.findByBoardIdAndWriterType(request.getId(), WriterEnum.ANONYMOUS_TYPE);
         if (boardTb.isPresent()) {
             if (!encoder.matches(request.getPassword(), boardTb.get().getPassword())) throw new BoardException(BoardResultCode.UN_MATCH_PASSWORD);
